% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_param.R
\name{as_param}
\alias{as_param}
\alias{txt}
\title{as_param}
\usage{
as_param(..., fun = NULL)

txt(x)
}
\arguments{
\item{...}{list of named and unnamed parameters}

\item{fun}{character:}

\item{x}{character: replaces \code{"x"} by \code{"'x'"}}
}
\value{
a character as parameter list of function call
}
\description{
Create a parameter list or a function call. For a function call \code{fun} must be explicitly given.
}
\examples{
as_param(letters[1:5])
as_param(txt(letters[1:5]))
as_param(a=txt("a"))
as_param(txt(letters[1:5]), fun="c")
}
