% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagplot2.R
\name{bagplot2}
\alias{bagplot2}
\title{bagplot}
\usage{
bagplot2(
  x,
  y = NULL,
  colorbag = NULL,
  colorloop = NULL,
  colorchull = NULL,
  databag = TRUE,
  dataloop = TRUE,
  plot.fence = FALSE,
  type = "hdepth",
  sizesubset = 500,
  extra.directions = FALSE,
  options = NULL,
  ...
)
}
\arguments{
\item{x, y}{the x and y arguments provide the x and y coordinates for the bagplot.
Any reasonable way of defining the coordinates is acceptable. See the function \code{xy.coords}
for details. If supplied separately, they must be of the same length.}

\item{colorbag}{The color of the bag (which contains the 50\% observations with largest depth).}

\item{colorloop}{The color of the loop (which contains the regular observations).}

\item{colorchull}{When the bagplot is based on halfspace depth, the depth region with maximal depth is plotted. This argument controls its color.}

\item{databag}{Logical indicating whether data points inside the bag need to be plotted. \cr
  Defaults to \code{TRUE}.}

\item{dataloop}{Logical indicating whether data points inside the fence need to be plotted. \cr
  Defaults to \code{TRUE}.}

\item{plot.fence}{Logical indicating whether the fence should be plotted. \cr
  Defaults to \code{FALSE}.}

\item{type}{Determines the depth function used to construct the bagplot: \code{"hdepth"} for halfspace depth, \code{"projdepth"} for projection depth and \code{"sprojdepth"} for skewness-adjusted projection depth. \cr
              Defaults to \code{"hdepth"}.
}

\item{sizesubset}{When computing the bagplot based on halfspace depth,
                    the size of the subset used to perform the main
                    computations. See Details for more information. \cr
                    Defaults to \eqn{500}.
                    }

\item{extra.directions}{Logical indicating whether additional directions should
                         be considered in the computation of the fence for the
                         bagplot based on projection depth or skewness-adjusted
                         projection depth. If set to \code{TRUE} an additional 
                         250 equispaced directions
                         are added to the directions defined by the points in
                         \code{x} themselves and the center. If \code{FALSE}
                         only directions determined by the points in \code{x} are
                         considered. \cr
                         Defaults to \code{FALSE}.
                        }

\item{options}{A list of options to pass to the \code{projdepth} or \code{sprojdepth} function.
                 In addition the following option may be specified:
                  \itemize{
                      \item{ \code{max.iter} \cr
                            The maximum number of iterations in the bisection algorithm used
                            to compute the depth contour corresponding to the cutoff. See
                            \code{depthContour} for more information. \cr
                            Defaults to \eqn{100}.
                           }
                          }

                }

\item{...}{further parameters given to \code{plot}}
}
\value{
Invisibly the result of the call to \link[mrfDepth:compBagplot]{mrfDepth::compBagplot}
}
\description{
A non-ggplot2 bagplot based on \link[mrfDepth:bagplot]{mrfDepth::bagplot}.
}
\details{
The bagplot has been proposed by Rousseeuw et al. (1999) as a generalisation of the
boxplot to bivariate data. It is constructed based on halfspace depth and as such is invariant
under affine transformations. Similar graphical representations can be obtained by means of other
depth functions, as illustrated in Hubert and Van der Veeken (2008) and in Hubert et al. (2015).
See \link[mrfDepth:compBagplot]{mrfDepth::compBagplot} for more details.

The deepest point is indicated with a "*" sign, the outlying observations with red points.
}
\examples{
bagplot2(iris$Sepal.Length, iris$Sepal.Width)
bagplot2(iris[,1:2])
bagplot2(iris[,3:4], title="Bagplot with Tukey depth", xlab="Petal.Length", ylab="Petal.Width") 
#
library("mrfDepth")
data("bloodfat")
result <- compBagplot(bloodfat)
bagplot(result, colorbag = rgb(0.2,0.2,0.2), colorloop = "green")
}
\references{
Rousseeuw P.J., Ruts I., Tukey, J.W. (1999). The bagplot: a bivariate boxplot. The American Statistician, 53, 382–387.

Hubert M., Van der Veeken S. (2008). Outlier detection for skewed data. Journal of Chemometrics, 22, 235–246.

Hubert M., Rousseeuw P.J., Segaert, P. (2015). Rejoinder to 'Multivariate functional outlier detection'. Statistical Methods & Applications, 24, 269–277.
}
\seealso{
\link[mrfDepth:compBagplot]{mrfDepth::compBagplot} and \link[mrfDepth:bagplot]{mrfDepth::bagplot}
}
