% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binData.R
\name{binData}
\alias{binData}
\title{binData}
\usage{
binData(
  data,
  bins,
  out = c("data.frame", "table", "binned"),
  val = c("mid", "interval"),
  pretty = TRUE,
  numeric = TRUE
)
}
\arguments{
\item{data}{object: a data.frame or object that can be converted to a data frame with variables to bin}

\item{bins}{integer: number of bins, will be recycled if necessary}

\item{out}{character: output type, either \code{"data.frame"}, \code{"table"}, or \code{"binned"}}

\item{val}{character: values for outer, eiter \code{"mids"} (interval centers), or \verb{"interval} (interval number)}

\item{pretty}{logical: should be \link[base:pretty]{base::pretty} used or minimum and maximum (default: \code{TRUE})}

\item{numeric}{logical: return output a \code{factor} or as \code{numeric} (default: \code{TRUE})}
}
\value{
a data frame or table with the results
}
\description{
Bins each variable in \code{data} in \code{bins} Bins. It can return a data frame (\code{out="data.frame"}), a table with the counts (\code{out="table"}), or
a table converted to a data frame with an additional variable \code{Freq}. The values can be either the bin mids (\code{val="mids"}) or
the bin numbers (\code{val="interval"}). If possible all variables contain an attribute \code{breaks} with breaks used.
}
\examples{
df <- data.frame(x=runif(25), y=runif(25))
binData(df, 5, 'data.frame')
binData(df, 5, 'table')
binData(df, 5, 'binned')
}
