% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_data.R
\name{character_data}
\alias{character_data}
\title{character_data}
\usage{
character_data(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  na.action = "NA",
  ...,
  title = NULL
)
}
\arguments{
\item{x}{vector, matrix or data frame}

\item{select}{vector: indicating columns to select (default: \code{NULL})}

\item{out}{output as \code{data.frame} (default), \code{matrix}, or \code{vector}}

\item{na.action}{function or character: indicates what should happen when the data contain NAs}

\item{...}{unused}

\item{title}{character: title attribute (default \code{NULL})}
}
\value{
the desired R object
}
\description{
Converts a matrix or data frame into a character vector, matrix or data frame. If \code{na.action} is a character then
all \code{NA}s are replaced by \code{na.action} (default: \code{na.action="NA"}). If \code{na.action} is a function then the function will be
applied to the result.
}
\examples{
character_data(iris)
character_data(iris, out="matrix")
character_data(iris, out="vector")
}
