% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackages.R
\name{checkPackages}
\alias{checkPackages}
\alias{installPackages}
\title{checkPackages, installPackages}
\usage{
checkPackages(
  ...,
  plotmodule = NULL,
  add = c("tools", "devtools", "formatR", "highlight", "shiny", "shinydashboard",
    "shinydashboardPlus", "shinyWidgets", "DT", "sortable", "base64enc"),
  error = FALSE
)

installPackages(
  plotmodule = NULL,
  add = c("tools", "devtools", "formatR", "highlight", "shiny", "shinydashboard",
    "shinydashboardPlus", "shinyWidgets", "DT", "sortable", "base64enc")
)
}
\arguments{
\item{...}{character: name(s) of package}

\item{plotmodule}{character: name(s) of plot modules to check for packages}

\item{add}{character: names of default packages to check (default: \code{c("highlight", "formatR", "shiny", "shinydashboard", "shinydashboardPlus", "DT")})}

\item{error}{logical: should a error thrown if one or more package are missing? (default: \code{FALSE})}
}
\value{
\code{TRUE} if successful otherweise an error will be thrown
}
\description{
Checks if a package is installed without loading it. Returns a logical vector with \code{TRUE} or \code{FALSE} for each package checked.
}
\examples{
checkPackages("graphics", add=NULL)          # checks if 'graphics' is installed
if (interactive()) checkPackages("graphics") # checks if 'graphics', 'shiny', ... are installed
if (interactive()) installPackages()         # installs all packages to show ALL plots
}
