% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_data.R
\name{color_data}
\alias{color_data}
\title{color_data}
\usage{
color_data(x, colpal = grDevices::hcl.colors, select = NULL, ..., title = NULL)
}
\arguments{
\item{x}{vector, matrix, or data frame}

\item{colpal}{color palette (default: \link[grDevices:palettes]{grDevices::hcl.colors})}

\item{select}{vector: indicating columns to select (default: \code{1})}

\item{...}{further parameters to \link{group_data}}

\item{title}{character: title attribute (default \code{NULL})}
}
\value{
a color vector
}
\description{
Assigns a color to the data \code{x} based on the color palette \code{colpal}.
}
\examples{
color_data(iris)
color_data(as.matrix(iris))
color_data(iris$Species)
}
