% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_hclust.R
\name{color_hclust}
\alias{color_hclust}
\title{color_hclust}
\usage{
color_hclust(
  x,
  normalize = 1,
  ncol = 2,
  colpal = grDevices::hcl.colors,
  dist = "euclidean",
  na.action = stats::na.pass,
  ...
)
}
\arguments{
\item{x}{a numeric matrix, data frame or "dist" object.}

\item{normalize}{integer: normalization method (default: \code{1})
\itemize{
\item 0: no rescaling
\item 1: \eqn{(x-min(x))/(max(x)-min(x))}
\item 2: \eqn{(x-mean(x))/sd(x)}
}}

\item{ncol}{integer: maximal number colors}

\item{colpal}{color palette: a function which generates "ncol" colors with "colpal(ncol)" (default: \link[grDevices:palettes]{grDevices::hcl.colors})}

\item{dist}{the distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra" or "binary"(default: \code{euclidean})}

\item{na.action}{a function which indicates what should happen when the data contain NAs (default: \code{na.pass})}

\item{...}{further parameters given to \link[stats:hclust]{stats::hclust}}
}
\value{
a color vector
}
\description{
Determines colors for \code{x} based on \link[stats:hclust]{stats::hclust}. \code{x} is normalized according \link{normalize}.
}
\examples{
color_hclust(iris[,-5], ncol=6)
}
