% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTo.R
\name{convertTo}
\alias{convertTo}
\title{convertTo}
\usage{
convertTo(x, coln, rown, title, out = c("data.frame", "matrix", "vector"))
}
\arguments{
\item{x}{vector, matrix, or data frame: input}

\item{coln}{character: column names if possible}

\item{rown}{character: row names if possible}

\item{title}{character: for title attribute}

\item{out}{character: either}
}
\value{
the desired output object
}
\description{
Converts an input object (vector, matrix or data frame) to an output asccording to the format in \code{out}.
Variable, Row and column names are set, if possible as well as a attribute \code{title}.
}
\examples{
str(convertTo(pi, "Col1", "Row1", "Title", out='data.frame'))
str(convertTo(pi, "Col1", "Row1", "Title", out='matrix'))
str(convertTo(pi, "Col1", "Row1", "Title", out='vector'))
}
