% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_data.R
\name{factor_data}
\alias{factor_data}
\title{factor_data}
\usage{
factor_data(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  exclude = NULL,
  na.action = stats::na.pass,
  ...,
  title = NULL
)
}
\arguments{
\item{x}{vector, matrix, or data frame}

\item{select}{vector: indicating columns to select (default: \code{NULL})}

\item{out}{output as \code{data.frame} (default), \code{matrix}, or \code{vector}}

\item{exclude}{vector: values to be excluded when forming the set of levels (default: \code{NULL})}

\item{na.action}{a function which indicates what should happen when the data contain NAs (default: \link[stats:na.fail]{stats::na.pass})}

\item{...}{further parameters to \link{character_data}}

\item{title}{character: title attribute (default \code{NULL})}
}
\value{
a one-column matrix with the merged groups
}
\description{
Creates a single group variable from the data \code{x}.
}
\examples{
factor_data(iris$Species, out="vector")
factor_data(iris)
}
