% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatCommands.R
\name{formatCommands}
\alias{formatCommands}
\title{formatCommands}
\usage{
formatCommands(cmds)
}
\arguments{
\item{cmds}{characater: R code}
}
\value{
HTML code for the splot app
}
\description{
formatCommands
}
\examples{
formatCommands('print("Hello World!")')
}
