% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModules.R
\name{getModules}
\alias{getModules}
\title{getModules}
\usage{
getModules(pattern, path = getShinyOption("smvgraph.path"))
}
\arguments{
\item{pattern}{character: character string containing a regular expression, currently are used \code{plot_*.R} and \code{color_*.R}}

\item{path}{character: containing the path where to search the modules}
}
\value{
a list with the modules
}
\description{
Returns a list of available module as a list.
}
\examples{
library("shiny")
getModules('plot_*.R')   # get plots
getModules('color_*.R')  # get colors
}
