% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableInfo.R
\name{getVariableInfo}
\alias{getVariableInfo}
\title{getVariableInfo}
\usage{
getVariableInfo(data, n = 47)
}
\arguments{
\item{data}{data frame: input data set}

\item{n}{integer: character length for \code{values} (default: \code{47})}
}
\value{
a data frame with information about the variables of the input data set
}
\description{
Returns a data frame with one row for each variable in \code{data}:
}
\details{
\itemize{
\item \code{class} the \link[base:class]{base::class} of the variable
\item \code{missing} the number of missing values
\item \code{infinite} the number of infinite values
\item \code{unique} the number of unique values
\item \code{valid} the number of unique valid values (see \link{valid})
\item \code{values} the values with the decreasing frequency
}
}
\examples{
getVariableInfo(iris)
}
