% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableNames.R
\name{getVariableNames}
\alias{getVariableNames}
\title{getVariableNames}
\usage{
getVariableNames(x, xvar = NULL, num = TRUE)
}
\arguments{
\item{x}{data frame/matrix: data set to analyse}

\item{xvar}{character: variable names to analyse (default: \code{character(0)} = all variables)}

\item{num}{logical: should numerical or non-numerical variable use (default: \code{TRUE})}
}
\value{
character vector with variable names
}
\description{
Extracts variable names from a data frame or matrix (column names).
}
\examples{
getVariableNames(iris)
getVariableNames(iris, num=FALSE)
getVariableNames(normalize(iris, 0))
getVariableNames(normalize(iris, 0), num=FALSE)
}
