% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_data}
\alias{group_data}
\alias{group_data.default}
\alias{group_data.matrix}
\alias{group_data.data.frame}
\title{group_data}
\usage{
group_data(x, ...)

\method{group_data}{default}(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  ...,
  title = NULL
)

\method{group_data}{matrix}(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  ...,
  title = NULL
)

\method{group_data}{data.frame}(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  ...,
  title = NULL
)
}
\arguments{
\item{x}{vector, matrix, or data frame}

\item{...}{further parameters to \link{character_data}}

\item{select}{vector: indicating columns to select (default: \code{NULL})}

\item{out}{output as \code{data.frame} (default), \code{matrix}, or \code{vector}}

\item{title}{character: title attribute (default \code{NULL})}
}
\value{
a one-column matrix with the merged groups
}
\description{
Creates a single group variable from the data \code{x}.
}
\examples{
group_data(iris$Species, out="vector")
group_data(iris)
}
