% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_min.R
\name{jitter_min}
\alias{jitter_min}
\title{jitter_min}
\usage{
jitter_min(x, factor = 1, amount = 0)
}
\arguments{
\item{x}{numeric: vector to which jitter should be added}

\item{factor}{numeric: multiplier for \code{amount} (default: \code{1})}

\item{amount}{numeric: amount for jittering (default: \code{0})}
}
\value{
jittered data
}
\description{
Add a small amount of noise to a numeric vector. The result is \code{x + runif(n, -a, a)} where \code{n <- length(x)}
and \code{a <- abs(factor*amount)} argument. If \code{amount==0} then \code{amount}
is set to \code{1e-6} times the smallest non-zero distance between adjacent unique \code{x} values.
In case of no non-zero distances \code{amount} is set to \code{1e-6*(1+min(abs(x)))}.
Note that \code{jitter_min} delivers different results then \link[base:jitter]{base::jitter}.
}
\examples{
jitter_min(runif(6))
jitter_min(rep(0, 7))
jitter_min(rep(10000, 5))
}
