% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggit.R
\name{loggit}
\alias{loggit}
\alias{stop}
\alias{warning}
\alias{message}
\alias{read_logs}
\alias{set_logfile}
\title{loggit}
\usage{
loggit(log_lvl, log_msg)

read_logs()

set_logfile()
}
\arguments{
\item{log_lvl}{character: Level of log output. In actual practice, one of "DEBUG", "INFO", "WARN", and "ERROR" are common, but any string may be supplied}

\item{log_msg}{character: Main log message}
}
\value{
Nothing.
}
\description{
Stores in a temporary file the log messages including messages, warnings and errors.
}
\examples{
if (interactive()) {
  set_logfile()      # create a temporary file for logging
  loggit("DEBUG", "Hello world")
  read_logs()        # get a data frame with the current messages.
}
}
