% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_data.R
\name{numeric_data}
\alias{numeric_data}
\alias{numeric_data.default}
\alias{numeric_data.matrix}
\alias{numeric_data.data.frame}
\title{numeric_data

Converts a vector, matrix or data frame into a numeric vector, matrix or data frame.}
\usage{
numeric_data(x, ...)

\method{numeric_data}{default}(
  x,
  out = c("data.frame", "matrix", "vector"),
  na.action = stats::na.pass,
  ...,
  title = NULL
)

\method{numeric_data}{matrix}(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  na.action = stats::na.pass,
  ...,
  title = NULL
)

\method{numeric_data}{data.frame}(
  x,
  select = NULL,
  out = c("data.frame", "matrix", "vector"),
  na.action = stats::na.pass,
  ...,
  title = NULL
)
}
\arguments{
\item{x}{vector, matrix or data frame}

\item{...}{unused}

\item{out}{output as \code{data.frame} (default), \code{matrix}, or \code{vector}}

\item{na.action}{a function which indicates what should happen when the data contain NAs (default: \link[stats:na.fail]{stats::na.pass})}

\item{title}{character: title attribute (default \code{NULL})}

\item{select}{vector: indicating columns to select (default: \code{NULL})}
}
\value{
the desired R object
}
\description{
numeric_data

Converts a vector, matrix or data frame into a numeric vector, matrix or data frame.
}
\examples{
numeric_data(iris)
numeric_data(iris, out="matrix")
numeric_data(iris, out="vector")
}
