% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_parcoord.R
\name{order_parcoord}
\alias{order_parcoord}
\title{order_parcoord}
\usage{
order_parcoord(x, method = "spearman", ...)
}
\arguments{
\item{x}{data matrix}

\item{method}{numeric: order method (default: \code{"spearman"})}

\item{...}{further parameters given to \code{\link[stats:cor]{stats::cor()}}}
}
\value{
order of column vectors
}
\description{
Returns a reordering of the columns of \code{x} to visualize highly correlated variable pairs
based on a cluster analysis of the correlation matrix.
If no colum names are given then \code{V1}, \code{V2}, ... will be used.
}
\examples{
order_parcoord(iris)
}
