% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(data, title, shorten = TRUE)
}
\arguments{
\item{data}{matrix or data.frame}

\item{title}{character: prefix title for columns if they do not exist}

\item{shorten}{integer: should the title be shortened (default: \code{TRUE})}
}
\value{
matrix
}
\description{
Prepares a matrix or data frame:
}
\details{
\itemize{
\item extracts all numeric variables or columns,
\item deletes non-complete observations, and
\item creates column titles if they do not exist.
}
}
\examples{
prepare_data(iris[,-5], "iris[,-5]")
}
