% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{with_progress}
\alias{with_progress}
\alias{set_progress}
\alias{inc_progress}
\title{Progress}
\usage{
with_progress(...)

set_progress(...)

inc_progress(...)
}
\arguments{
\item{...}{see [\link[shiny:withProgress]{shiny::withProgress}]}
}
\value{
see [\link[shiny:withProgress]{shiny::withProgress}]
}
\description{
Reports progress to the user during long-running operations.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
   options(device.ask.default = FALSE)
   ui <- fluidPage(plotOutput("plot"))
   #
   server <- function(input, output) {
     output$plot <- renderPlot({
       with_progress(message = 'Calculation in progress',
                     detail = 'This may take a while...', value = 0, {
                     for (i in 1:15) {
                       inc_progress(1/15)
                       Sys.sleep(0.25)
                     }
                   })
       plot(cars)
   })
 }
 #
 shinyApp(ui, server)
}
}
