% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.R
\name{pyramid}
\alias{pyramid}
\title{pyramid}
\usage{
pyramid(
  tab,
  gap = 0,
  left = list(col = "red"),
  right = list(col = "blue"),
  ...
)
}
\arguments{
\item{tab}{table: a table with two columns}

\item{gap}{numeric(2): relative size of gap in \code{y}- and \code{x}-direction (default: \code{c(0,0)})}

\item{left}{list: parameters for the left polygons (default: \code{list(col="red")})}

\item{right}{list: parameters for the right polygons (default: \code{list(col="blue")})}

\item{...}{further parameters to use in \link[graphics:plot.default]{graphics::plot.default}}
}
\value{
a pyramid plot
}
\description{
pyramid
}
\examples{
data("Boston", package="MASS")
tab <- table(data.frame(Boston$rad, Boston$chas))
pyramid(tab, main="Absolute frequencies")
pyramid(tab, gap=c(0.2, 0.2))
rtab <- tab/sum(tab)
pyramid(rtab, gap=c(0.2, 0.2), main="Relative frequencies")
ctab <- proportions(tab, 2)
pyramid(ctab, gap=c(0.2, 0.2), main="Conditional frequencies on columns")
rtab <- proportions(tab, 1)
pyramid(rtab, gap=c(0.2, 0.2), main="Conditional frequencies on rows")
# zebraing 
pyramid(tab, gap=c(0.2, 0.2), 
        left=list(list(col="black"), list(col="white")), 
        right=list(list(col="blue"), list(col="green")))
}
