% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandrews.R
\name{sandrews}
\alias{sandrews}
\title{sandrews}
\usage{
sandrews(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{further parameters given to \code{\link[=andrews]{andrews()}}}
}
\value{
nothing
}
\description{
Shiny app for creating an Andrews curve diagram with interactive variable selection.
}
\examples{
if (interactive()) sandrews(normalize(iris))
}
