% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schernoff.R
\name{schernoff}
\alias{schernoff}
\title{schernoff}
\usage{
schernoff(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{further parameters given to \link[DescTools:PlotFaces]{DescTools::PlotFaces}}
}
\value{
nothing
}
\description{
Shiny app for creating a Chernoff faces plot with interactive variable selection.
}
\examples{
if (interactive()) schernoff(normalize(iris))
}
