% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbscan.R
\name{sdbscan}
\alias{sdbscan}
\title{sdbscan}
\usage{
sdbscan(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{...}{further parameters given to \link[dbscan:dbscan]{dbscan::dbscan}}
}
\value{
nothing
}
\description{
Shiny app which allows to run a cluster analysis with DBSCAN with interactive choice of variables,
core distance, and minimal neighbours.
}
\examples{
if (interactive()) sdbscan(iris)
}
