% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdistance.R
\name{sdistance}
\alias{sdistance}
\title{sdistance}
\usage{
sdistance(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app which shows the contribution of each variable to the distance between two observations
with interactive variable selection. If \eqn{dijk} is the distance between observations \eqn{i}
and \eqn{j} in variable \eqn{k} then the contribution is computed:
}
\details{
\itemize{
\item Total variance: \eqn{var_k/sum(var_k)} with \eqn{var_k} the variance of the \eqn{k}th variable
\item Minimum: \eqn{dijk==min_k(dijk)}
\item Manhattan: \eqn{dijk/sum(dijk)}
\item Gower: \eqn{dijk} is rescaled to \eqn{[0, 1]} in each variable and then \eqn{dijk/sum(dijk)}
\item Euclidean: \eqn{dijk^2/sum(dijk^2)}
\item Manhattan: \eqn{dijk/sum(dijk)}
\item Maximum: \eqn{dijk==max_k(dijk)}
}
}
\examples{
if (interactive()) sdistance(iris)
}
