% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfactor.R
\name{sfactor}
\alias{sfactor}
\title{sfactor}
\usage{
sfactor(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app for doing a factor analysis with interactive variable selection.
}
\examples{
if (interactive()) sfactor(iris)
}
