% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shclust.R
\name{shclust}
\alias{shclust}
\title{shclust}
\usage{
shclust(data, xvar = character(0), colpal = rainbow, ..., color = "black")
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{colpal}{color palette for cluster colors (default: \link[grDevices:palettes]{grDevices::rainbow})}

\item{...}{unused}

\item{color}{color: used for dendrogram (default: \code{"black"})}
}
\value{
nothing
}
\description{
Shiny app which allows to run a hierarchical cluster analysis with interactive choice of variables,
distance, and agglomeration method.
}
\examples{
if (interactive()) shclust(iris, col=rainbow)
}
