% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smclust.R
\name{smclust}
\alias{smclust}
\title{smclust}
\usage{
smclust(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app which allows to run a EM clustering with interactive choice of variables.
}
\examples{
if (interactive()) smclust(iris)
}
