% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smosaic.R
\name{smosaic}
\alias{smosaic}
\title{smosaic}
\usage{
smosaic(data, xvar = character(0), yvar = character(0), ...)
}
\arguments{
\item{data}{table or data.frame}

\item{xvar}{character: names of selected variables for x-axis}

\item{yvar}{character: names of selected variables for y-axis}

\item{...}{further parameters given to \link[graphics:mosaicplot]{graphics::mosaicplot}}
}
\value{
nothing
}
\description{
Shiny app for creating a Mosaic plot with interactive variable selection.
}
\examples{
if (interactive()) smosaic(Titanic)
dfTitanic <- toDataframe(Titanic)
if (interactive()) smosaic(dfTitanic)
}
