% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortbin.R
\name{sortbin}
\alias{sortbin}
\title{sortbin}
\usage{
sortbin(
  x,
  sortCol = 1,
  decreasing = FALSE,
  na.last = TRUE,
  nBins = NA,
  equibin = TRUE
)
}
\arguments{
\item{x}{data frame}

\item{sortCol}{numeric/character: names or indices of variable used for sorting (default: \code{1})}

\item{decreasing}{logical: should the sort order be increasing or decreasing (default: \code{FALSE})}

\item{na.last}{logical: for controlling the treatment of NAs (default: \code{TRUE})}

\item{nBins}{integer: maximal number of bins (default: \code{NA}).}

\item{equibin}{logical: should the number of observations equal per bin (default: \code{TRUE})}
}
\value{
(non-sequential) bin numbers as integer
}
\description{
Sorts and bins the rows of the data frame \code{x} according the sorting columns in \code{sortCol}.
\code{decreasing} and \code{na.last} are recycled is necessary. If \code{equibin} is \code{TRUE}
and \code{nBins==NA} then \code{nBins} is set to \code{100}.  If \code{equibin} is \code{FALSE}
and \code{nBins==NA} then the bins are returned as they come from sorting; only identical values are
in one bin. If \code{nBins} is positive then the bins are merged until \code{nBins} reached.
Note that the numbers of observations per bin may vary.
}
\examples{
data("Boston", package="MASS")
tableplot(Boston, bin=sortbin(Boston))
}
