% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spairs.R
\name{spairs}
\alias{spairs}
\title{spairs}
\usage{
spairs(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{further parameters given to \code{\link[graphics:pairs]{graphics::pairs()}}}
}
\value{
nothing
}
\description{
Shiny app for creating a scatterplot matrix with interactive variable selection.
}
\examples{
if (interactive()) spairs(iris, col=as.factor(iris$Species))
}
