% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparcoord.R
\name{sparcoord}
\alias{sparcoord}
\title{sandrews}
\usage{
sparcoord(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{further parameters given to \link[MASS:parcoord]{MASS::parcoord}}
}
\value{
nothing
}
\description{
Shiny app for creating a Parallel Coordinate plot with interactive variable selection.
}
\examples{
if (interactive()) sparcoord(iris)
}
