% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sradar.R
\name{sradar}
\alias{sradar}
\title{sradar}
\usage{
sradar(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the plot}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app for creating radar charts with interactive variable selection.
}
\examples{
if (interactive()) sradar(iris)
}
