% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableplot.R
\name{tableplot}
\alias{tableplot}
\title{tableplot}
\usage{
tableplot(
  x,
  select = NULL,
  subset = NULL,
  bin = NULL,
  yj = NA,
  IQR_bias = 5,
  colpal = grDevices::rainbow,
  color.NA_num = "gray75",
  color.NA = "grey75",
  color.num = "lightblue",
  color.box = "deepskyblue",
  color.line = "black",
  box.lower = NULL,
  box.upper = NULL,
  box.line = NULL,
  cex.main = 1,
  cex.legend = 1,
  width = 1,
  height = 0.15
)
}
\arguments{
\item{x}{data frame}

\item{select}{numeric/character: variable to show in the plot (default: \code{NULL})}

\item{subset}{numeric: index of observations to show}

\item{bin}{integer: bin numbers to which a observations belongs (default: \code{NULL} = all)}

\item{yj}{numeric: Yeo Johnson coefficient (default: \code{NA}). If \code{NA} then it will be set to 0 (=log) or 1 (=identity)}

\item{IQR_bias}{numeric: parameter that determines when a logarithmic scale is used when \code{yj} is set to \code{NA}. The argument IQR_bias is multiplied by the interquartile range as a test.}

\item{colpal}{color palette to draw (default: \code{rainbow})}

\item{color.NA_num}{color for missing of infinity values for numeric variables (default: \code{gray75})}

\item{color.NA}{color for missing values for categorical variables (default: \code{grey75})}

\item{color.num}{color for lower box for numeric variables (default: \code{lightblue})}

\item{color.box}{color for upper box for numeric variables (default: \code{deepskyblue})}

\item{color.line}{color for line in upper box for numeric variables  (default: \code{black})}

\item{box.lower}{function: determine lower border in upper box for numeric variables (default: \code{NULL}). If \code{NULL} then \code{min(.,na.rm=TRUE)} is used.}

\item{box.upper}{function: determine upper border in upper box for numeric variables (default: \code{NULL}). If \code{NULL} then \code{max(.,na.rm=TRUE)} is used.}

\item{box.line}{function: determine line position in upper box for numeric variables (default: \code{NULL}). If \code{NULL} then \code{mean(.,na.rm=TRUE)} is used.}

\item{cex.main}{number: magnification to be used for the titles (default: \code{1})}

\item{cex.legend}{number: magnification to be used for the legends (default: \code{1})}

\item{width}{number: width of percentage axis (default: \code{1}). If \code{1} then the width is as wide as a plot.}

\item{height}{number: percentage of the height of the legends (default: \code{0.15})}
}
\value{
nothing
}
\description{
A tableplot is a visualisation of multivariate data sets. Each column represents a variable
and each row bin is an aggregate of a certain number of records. For numeric variables, a value box is
plotted with minimum, mean (black line) and maximum value. If any missing values in a bin of
a numeric variable appear the box left from the value box is plotted in gray.
For categorical variables, a stacked bar chart is depicted of the proportions of categories.
Missing values are taken into account.
}
\details{
The idea and some code of the tableplot is taken from \href{https://github.com/mtennekes/tabplot}{tableplot package}
by Martijn Tennekes and Edwin de Jonge.
It differs from their package by
\itemize{
\item multicolumn sorting is possible, and
\item no support for 'ff' (out of memory vectors).
}
}
\examples{
data("Boston", package="MASS")
tableplot(Boston, bin=sortbin(Boston))
}
\references{
Tennekes, M., Jonge, E. de, Daas, P.J.H. (2013) \href{https://www.jds-online.com/files/JDS-1108.pdf}{Visualizing and Inspecting Large Datasets with Tableplots}, Journal of Data Science 11 (1), 43-58.
}
