% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_style.R
\name{tags_style}
\alias{tags_style}
\title{tags_style}
\usage{
tags_style()
}
\value{
HTML style
}
\description{
A default style for the Shiny apps:
\itemize{
\item \code{HTML(".black-text .rank-list-item { color: #000000; }")},
\item \code{HTML(".shiny-split-layout > div { overflow: visible;}")}, and
\item \code{HTML(".main-sidebar { font-size: 10px; }")}.
}
}
\examples{
tags_style()
}
