% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{template}
\usage{
template(text, ...)
}
\arguments{
\item{text}{a code template}

\item{...}{further parameters}
}
\value{
a character vector
}
\description{
Each line of a code template consists of condition based on the unnamed parameters and R code in which replacements
with named parameters done.
}
\examples{
template("
1:   'Hello {{letter}}'
!1:  'Good-bye {{letter}}'
         ",
         letter=sample(LETTERS, 1),
         runif(1)<0.5 #1 = first unnamed parameter
         )
}
