% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test data}
\format{
A data frame with \code{n=25} rows and 8 variables:
\describe{
\item{xu}{\code{runif(n)} with a \code{NA}, \code{NaN}, \code{Inf}, \code{-Inf}}
\item{xn}{\code{rnorm(n, 0, 2)} with a \code{NA}, \code{NaN}}
\item{x0}{\code{rep(0, n)}}
\item{xi}{\verb{as.integer(rnorm(n, 0, 2)} with a \code{NA}, \code{NaN}}
\item{x2}{\code{sample(c(0,1), size=n, replace=TRUE)}}
\item{gf}{\verb{factor(as.integer(rnorm(n, 0, 2))} with a \code{NA}}
\item{go}{\verb{ordered(as.integer(rnorm(n, 0, 2))} with a new level \code{10}}
\item{gn}{\verb{ordered(as.integer(rnorm(n, 0, 2))} with a \code{NA}}
\item{gc}{\verb{as.character(as.integer(rnorm(n, 0, 2))} with a \code{NA} and \code{""}}
\item{gl}{\verb{sample(c(T,F), size=n, replace=TRUE))} with a \code{NA}}
\item{g0}{\verb{rep("constant, n)}}
\item{g2}{\code{sample(c(T,F), size=n, replace=TRUE)}}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
testdata
}
\description{
A data frame containing various variable types and special values.
}
\keyword{datasets}
