% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toChoice.R
\name{toChoice}
\alias{toChoice}
\title{toChoice}
\usage{
toChoice(method = NA, ...)
}
\arguments{
\item{method}{integer: which method is used for creating the list elements}

\item{...}{character: choice values}
}
\value{
a list
}
\description{
The elements in \code{...} will coerced into one text vector. The entries will either the text (\code{method==NA}) or integer number starting at \code{method}.
The first letter of the list element names will be capitalized.
}
\examples{
txt <- c("the", "quick", "brown", "fox", "jumps", "over", "the", "lazy", "dog")
toChoice(NA, txt)
toChoice(0, txt) # integer sequence starts at zero
}
