% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toDataframe.R
\name{toDataframe}
\alias{toDataframe}
\title{toDataframe}
\usage{
toDataframe(obj, name = NULL, ...)
}
\arguments{
\item{obj}{R object (\code{table} or \code{ts}) to convert to a data frame}

\item{name}{character: vector of variable name(s), only use for a \code{ts} object}

\item{...}{further parameters given to \link[base:table]{base::as.data.frame.table}}
}
\value{
a data frame
}
\description{
Converts a table to a full data frame.
}
\examples{
toDataframe(Titanic)
toDataframe(austres)
}
