% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLayout.R
\name{toLayout}
\alias{toLayout}
\title{toLayout}
\usage{
toLayout(len, sel = 0, unused = 0)
}
\arguments{
\item{len}{integer: minimal size of layout}

\item{sel}{integer: select less or more rows (default: \code{0})}

\item{unused}{integer: number of unused entries (default: \code{0})}
}
\value{
a matrix
}
\description{
Given a (minimal) length \code{len} and the number unused entries a layout is generated.
If \code{sel<0} then less rows and more columns are used and if \code{sel>0} then more rows and less columns are used.
}
\examples{
toLayout(13)
}
