% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_season.R
\name{trend_season}
\alias{trend_season}
\alias{trend_season.default}
\alias{print.trend_season}
\alias{summary.trend_season}
\alias{plot.trend_season}
\title{Trend and aeasonality estimation of a univariate time series}
\usage{
trend_season(t, ...)

\method{trend_season}{default}(
  t,
  trend = c("constant", "linear", "exponential"),
  season = c("none", "additive", "multiplicative"),
  ...
)

\method{print}{trend_season}(x, ...)

\method{summary}{trend_season}(object, ...)

\method{plot}{trend_season}(x, y, which = 1, ...)
}
\arguments{
\item{t}{ts: time series object}

\item{...}{unused}

\item{trend}{character: trend method, either \code{none} (default), \code{linear} or \code{exponential}}

\item{season}{character: seasonality method, either \code{none} (default), \code{additive} or \code{multiplicative}}

\item{x, object}{trend_season: estimated time series}

\item{y}{unused}

\item{which}{integer: what to plot, \code{1} time series and estimation (default) or \code{2} residuals}
}
\value{
\code{trend_season} returns a \code{trend_season} object with
\itemize{
\item \code{call} the function call
\item \code{ts} the input time series
\item \code{trend} the trend estimation (\code{ts} object)
\item \code{trend.residuals} the residuals of the trend estimation (\code{ts} object)
\item \code{season} the trend and season estimation (\code{ts} object)
\item \code{season.residuals} the residuals of the trend and season estimation (\code{ts} object)
\item \code{coefficients} the coefficients used in the seasonality estimation
\item \code{residuals} the residuals of the model
\item \code{fitted.values} the fitted values of the model
}
}
\description{
Estimate a trend and seasonaliyt for a time series.
Available functions:
\itemize{
\item \code{trend_season} to generate an estimate
\item \code{print} to print the estimate
\item \code{summary} to summarize the etsimate result
\item \code{plot} to plot the time series, its estimation and the residuals
\item \code{coef} to extract the coefficients if a seasonality estimation was done
\item \code{residuals} to extract the residuals of the model
\item \code{fitted} to the fitted values
}
}
\examples{
tts <- trend_season(austres, "linear")
print(tts)
summary(tts)
plot(tts)
plot(tts, which=2)
residuals(tts)
fitted(tts)
coef(tts)  # if NULL then no seasonality was estimated
}
