% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.R
\name{UIdatanormalization}
\alias{UIdatanormalization}
\alias{UIdistance}
\alias{UIobservations}
\alias{UImergegroups}
\alias{UIpointsize}
\alias{UIpointsymbol}
\alias{UIlinewidth}
\alias{UIlinetype}
\alias{UItextsize}
\alias{UIlegend}
\alias{UIlegendsize}
\alias{UIplottype}
\title{General UI elements}
\usage{
UIdatanormalization(
  sel = getShinyOption("smvgraph.current")$smvgraph_normalize
)

UIdistance(sel = getShinyOption("smvgraph.current")$smvgraph_distance)

UIobservations(n, sel = getShinyOption("smvgraph.current")$smvgraph_obs)

UImergegroups(n, sel = getShinyOption("smvgraph.current")$smvgraph_single)

UIpointsize(n, sel = getShinyOption("smvgraph.current")$smvgraph_cex)

UIpointsymbol(n, sel = getShinyOption("smvgraph.current")$smvgraph_pch)

UIlinewidth(n, sel = getShinyOption("smvgraph.current")$smvgraph_lwd)

UIlinetype(n, sel = getShinyOption("smvgraph.current")$smvgraph_lty)

UItextsize(n, sel = getShinyOption("smvgraph.current")$smvgraph_tex)

UIlegend(n, sel = getShinyOption("smvgraph.current")$smvgraph_legend)

UIlegendsize(n, sel = getShinyOption("smvgraph.current")$mvgraph_lex)

UIplottype(n, sel = getShinyOption("smvgraph.current")$smvgraph_type)
}
\arguments{
\item{sel}{selected element}

\item{n}{integer: number of observations}
}
\value{
an UI element for shiny
}
\description{
Some general UI elements for common use where last selected value is stored for reuse:
\itemize{
\item \code{UIplottype} plot type, defines \code{smvgraph_type}
\item \code{UIpointsymbol} plot symbol for point, defines \code{smvgraph_pch}
\item \code{UIpointsize} point size, defines \code{smvgraph_cex}
\item \code{UIlinetype} line type, defines \code{smvgraph_lty}
\item \code{UIlinewidth} line width, defines \code{smvgraph_lwd}
\item \code{UItextsize} text size, defines \code{smvgraph_tex}
\item \code{UIlegend} legend position, defines \code{smvgraph_legend}
\item \code{UIlegendsize} legend size, defines \code{smvgraph_lex}
\item \code{UIdatanormalization} should data be rescaled, defines \code{smvgraph_normalize} (no, minMax, mtandardization)
\item \code{UIdistance} distance to use, defines \code{smvgraph_distance}
\item \code{UIobservations} range of observations, defines \code{smvgraph_obs}
\item \code{UImergegroups} should a set of grouping variables merged into one group variable, defines \code{smvgraph_single}
}

From the top menu are the following input elements are defined
\itemize{
\item \code{input$smvgraph_pch} point symbol,
\item \code{input$smvgraph_cex} point size,
\item \code{input$smvgraph_lty} line type,
\item \code{input$smvgraph_lwd} line width,
\item \code{input$smvgraph_tex} text size, and
\item \code{inpus$smvgraph_legend} legend position.
}
}
\examples{
# none
}
