% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\alias{invalid}
\title{valid, invalid}
\usage{
valid(x, margin = 1:2, n = FALSE)

invalid(x, margin = 1:2, n = FALSE)
}
\arguments{
\item{x}{object: anything taht can be coerced to a data frame checked for valid/invalid values}

\item{margin}{integer: a vector giving the subscripts for which valid/invalid values looked for, e.g. \code{margin==1} indicates rows, \code{margin==2} indicates columns, otherwise indicates rows and columns.}

\item{n}{logical: should just the number of valid/invalid values returned or a logical matrix/vector}
}
\value{
a logical data frame, a logical vector or an integer
}
\description{
Computes the number a logical matrix or vector if all values are valid in \code{x} or each colum or row of \code{x}. Valid values for numeric variables are \code{is.finite(v)} and for other types \code{!is.na}

Computes the number a logical matrix or vector if any values are valid in \code{x} or each colum or row of \code{x}.
}
\examples{
data("testdata")
valid(testdata)             # matrix with logical entries if x has valid entry
valid(testdata, n=TRUE)     # number of valid entries in x
valid(testdata, 1)          # vector with logical entries if each row if x has valid entries
valid(testdata, 1, n=TRUE)  # number of rows with valid entries in x
valid(testdata$xu)
}
