% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_bucket_list.R
\name{variable_bucket_list}
\alias{variable_bucket_list}
\title{variable_bucket_list}
\usage{
variable_bucket_list(data, xvar, fun = NULL)
}
\arguments{
\item{data}{matrix: data}

\item{xvar}{character: names of variables to use}

\item{fun}{function: for returning a list of variable names (default: \code{NULL})}
}
\value{
character with variable names
}
\description{
Creates a \code{bucket_list} for a variable choice in Shiny app.
}
\examples{
# none yet
}
