% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shclust.R
\name{shclust}
\alias{shclust}
\title{shclust}
\usage{
shclust(data, xvar = character(0), ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app which allows to run a hierarchical cluster analysis with interactive choice of variables,
distance, and agglomeration method.
}
\examples{
if (interactive()) shclust(iris)
}
