% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skmeans.R
\name{skmeans}
\alias{skmeans}
\title{skmeans}
\usage{
skmeans(data, xvar = character(0), colpal = rainbow, ...)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{colpal}{color palette for cluster colors (default: \link[grDevices:palettes]{grDevices::rainbow})}

\item{...}{unused}
}
\value{
nothing
}
\description{
Shiny app which allows to run a k-means cluster analysis with interactive choice of variables.
}
\examples{
if (interactive()) skmeans(iris, col=rainbow)
}
