% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smclust.R
\name{smclust}
\alias{smclust}
\title{smclust}
\usage{
smclust(data, xvar = character(0), colpal = rainbow, ..., srt = 0)
}
\arguments{
\item{data}{matrix or data.frame}

\item{xvar}{character: names of selected variables for the clustering}

\item{colpal}{color palette for cluster colors (default: \link[grDevices:palettes]{grDevices::rainbow})}

\item{...}{unused}

\item{srt}{numeric: model string rotation in degrees (default: \code{0})}
}
\value{
nothing
}
\description{
Shiny app which allows to run a EM clustering with interactive choice of variables.
}
\examples{
if (interactive()) smclust(iris, colpal=rainbow)
}
