% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{splot}
\alias{splot}
\title{splot}
\usage{
splot(data, xvar = character(0), path = NULL)
}
\arguments{
\item{data}{data.frame: input data}

\item{xvar}{character: selected variables (default: \code{character(0)})}

\item{path}{character: path where to read the plot modules (default: \code{NULL})}
}
\value{
nothing
}
\description{
Shiny app for choosing a specific plot.
}
\examples{
if (interactive()) splot(iris)
}
